Move couse popup to different file
